backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
auth0-python==4.0.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
load-dotenv==0.1.0
multidict==6.0.4
openai==0.27.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==5.4.0
tqdm==4.65.0
twilio==7.17.0
urllib3==1.26.14
yarl==1.8.2
